team_code/requirements.txt
Flask==1.1.2
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-datastore==2.1.0
google-cloud-language==2.0.0
pandas
numpy
matplotlib

